fix: [lw-11830]: resolve pr comments
