+++ apps/browser-extension-wallet/src/views/browser-view/features/send-transaction/components/Form/LockedStakeRewardsBanner/LockedStakeRewardsBanner.module.scss
